+++ src/main/java/org/pda/etf/pdaetf/domain/etf/repository/query/EtfQueryRepository.java
@@ -8,8 +8,9 @@ import org.springframework.data.domain.Pageable;
 
 public interface EtfQueryRepository {
     Page<EtfRowDto> searchEtfs(String query, Long categoryId, Long currentUserId, Pageable pageable);
-    Page<EtfRowDto> findFavoriteEtfs(Long userId, String query, Pageable pageable);
+
     JPAQuery<EtfRowDto> buildBaseEtfQuery(BooleanExpression where,
                                                  BooleanExpression likedExpr,
                                                  boolean onlyFavorites);
+    void applySort(JPAQuery<?> query, Pageable pageable);
 }
